v 1.0

+++ posts/views.py
@@ -5,7 +5,7 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.views.decorators.cache import cache_page
 
 from .forms import PostForm, CommentForm
-from .models import Group, Post, User, Comment, Follow
+from .models import Group, Post, User, Follow
 
 
 @cache_page(60)
@@ -54,9 +54,10 @@ def profile(request, username):
     page_number = request.GET.get('page')
     page = paginator.get_page(page_number)
     following = (
-        request.user.is_authenticated
-        and author.following.filter(user=request.user).exists()
-    )
+            request.user.is_authenticated
+            and Follow.objects.filter(
+        user=request.user,
+        author=author).exists())
     return render(request, 'profile.html', {
         'page': page,
         'author': author,
@@ -67,12 +68,13 @@ def profile(request, username):
 
 def post_view(request, username, post_id):
     post = get_object_or_404(Post, author__username=username, id=post_id)
-    author = get_object_or_404(User, username=username)
-    comments = Comment.objects.filter(post=post)
+    comments = post.comments.all()
     following = (
-        request.user.is_authenticated
-        and author.following.filter(user=request.user).exists()
-    )
+            request.user.is_authenticated
+            and Follow.objects.filter(
+        user=request.user,
+        author=post.author).exists())
+    print(following)
     form = CommentForm()
     return render(request, 'post.html', {
         'post': post,
@@ -84,12 +86,12 @@ def post_view(request, username, post_id):
 
 
 @login_required
-def add_comment(request, username, post_id):
+def add_comment(request, username, post_id):  # комменты сохраняются не под юзером, а под автором поста
     post = get_object_or_404(Post, id=post_id,
                              author__username=username)
     form = CommentForm(request.POST or None)
     if not form.is_valid():
-        return redirect("posts:post", username=username, post_id=post_id)
+        return render(request, 'post.html', {'form': form})  # правильно ли это?
     comment = form.save(commit=False)
     comment.author = request.user
     comment.post = post
@@ -143,8 +145,10 @@ def profile_follow(request, username):
 def profile_unfollow(request, username):
     """Функция для отписки от автора"""
     author = get_object_or_404(User, username=username)
-    Follow.objects.filter(user=request.user, author=author).delete()
-    return redirect('posts:profile', username=username)
+    user_following_author = get_object_or_404(Follow, user=request.user, author=author)
+    if user_following_author:
+        Follow.objects.filter(user=request.user, author=author).delete()
+        return redirect('posts:profile', username=username)
 
 
 def page_not_found(request, exception):
